requirements.txt
confluent-kafka==2.2.0
cassandra-driver==3.25.0
hazelcast-python-client==5.5.0
faker==18.13.0
python-dateutil==2.8.2
pydantic==2.4.2
fastapi==0.103.1
uvicorn==0.23.2
aiokafka==0.8.1